resolve resize issues
