+++ web-app/src/components/Workspace/index.tsx
@@ -1,32 +1,22 @@
 import * as React from 'react'
 import { css, jsx } from '@emotion/core'
+import { useWindowResize } from '../../services/hooks/resize'
 
 interface Props {
   children: React.ReactElement
 }
 
-const resize = () => ({
-  width: window.innerWidth,
-  height: window.innerHeight,
-})
+const styles = {
+  page: {
+    display: 'flex' as 'flex',
+    position: 'relative' as 'relative',
+    margin: 0,
+    backgroundColor: 'white',
+  },
+}
 
 const Workspace = ({ children }: Props) => {
-  const [dimensions, setDimensions] = React.useState(resize())
-
-  // solution for windows getting off size
-  React.useEffect(() => {
-    setDimensions(resize())
-  }, [window.innerHeight, window.innerWidth])
-
-  const styles = {
-    page: {
-      display: 'flex' as 'flex',
-      position: 'relative' as 'relative',
-      margin: 0,
-      backgroundColor: 'white',
-    },
-  }
-
+  const dimensions = useWindowResize()
   return <div css={{ ...styles.page, ...dimensions }}>{children}</div>
 }
